Refactor FileSystem interface
